+++ src/controllers/common.go
@@ -4,8 +4,10 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
+	"runtime/debug"
 
 	"github.com/Parallels/prl-devops-service/basecontext"
+	"github.com/Parallels/prl-devops-service/errors"
 	"github.com/Parallels/prl-devops-service/models"
 )
 
@@ -21,15 +23,21 @@ func GetBaseContext(r *http.Request) *basecontext.BaseContext {
 
 func Recover(ctx basecontext.ApiContext, r *http.Request, w http.ResponseWriter) {
 	if err := recover(); err != nil {
-		ctx.LogErrorf("Recovered from panic: %v", err)
-		ReturnApiError(ctx, w, models.NewFromErrorWithCode(fmt.Errorf("Internal Server Error"), http.StatusInternalServerError))
+		ctx.LogErrorf("Recovered from panic: %v\n%v", err, debug.Stack())
+		sysErr := errors.NewWithCodef(http.StatusInternalServerError, "internal server error")
+		sysErr.NestedError = make([]errors.NestedError, 0)
+		sysErr.NestedError = append(sysErr.NestedError, errors.NestedError{
+			Message: fmt.Sprintf("%v", err.(error)),
+		}, errors.NestedError{
+			Message: string(debug.Stack()),
+		})
 
-		fmt.Printf("Recovered from panic: %v", err)
+		ReturnApiError(ctx, w, models.NewFromErrorWithCode(sysErr, http.StatusInternalServerError))
 	}
 }
 
 func ReturnApiError(ctx basecontext.ApiContext, w http.ResponseWriter, err models.ApiErrorResponse) {
-	ctx.LogErrorf("Error: %s", err.Message)
+	ctx.LogErrorf("Error: %v", err.Message)
 	w.WriteHeader(err.Code)
 
 	_ = json.NewEncoder(w).Encode(err)