+++ src/errors/nested_error.go
@@ -1,0 +1,8 @@
+package errors
+
+type NestedError struct {
+	Message     string
+	Path        string
+	Code        int
+	Description string
+}