+++ src/models/api_error_response.go
@@ -1,10 +1,20 @@
 package models
 
-import "github.com/Parallels/prl-devops-service/errors"
+import (
+	"github.com/Parallels/prl-devops-service/errors"
+)
+
+type ApiErrorStack struct {
+	Message     string `json:"message"`
+	Description string `json:"description,omitempty"`
+	Path        string `json:"path,omitempty"`
+	Code        int    `json:"code"`
+}
 
 type ApiErrorResponse struct {
-	Message string `json:"message"`
-	Code    int    `json:"code"`
+	Message string          `json:"message"`
+	Stack   []ApiErrorStack `json:"stack,omitempty"`
+	Code    int             `json:"code"`
 }
 
 func IsSystemError(err error) bool {
@@ -22,7 +32,8 @@ func GetSystemErrorCode(err error) int {
 	if !ok {
 		_, ok = err.(errors.SystemError)
 		if ok {
-			code := err.(errors.SystemError).Code()
+			sysError := err.(errors.SystemError)
+			code := sysError.Code()
 			if code == 0 {
 				return 400
 			}
@@ -41,16 +52,66 @@ func GetSystemErrorCode(err error) int {
 
 func NewFromError(err error) ApiErrorResponse {
 	if IsSystemError(err) {
+
 		code := GetSystemErrorCode(err)
-		return NewFromErrorWithCode(err, code)
+		result := ApiErrorResponse{
+			Message: err.Error(),
+			Code:    code,
+		}
+		return result
 	} else {
 		return NewFromErrorWithCode(err, 404)
 	}
 }
 
 func NewFromErrorWithCode(err error, code int) ApiErrorResponse {
-	return ApiErrorResponse{
-		Message: err.Error(),
-		Code:    code,
+	message := ApiErrorResponse{
+		Code: code,
 	}
+	if IsSystemError(err) {
+		sysError, ok := err.(errors.SystemError)
+		if ok {
+			message.Message = sysError.Message()
+			if len(sysError.NestedError) > 0 {
+				for _, nestedError := range sysError.NestedError {
+					stack := ApiErrorStack{
+						Message: nestedError.Message,
+					}
+					if nestedError.Path != "" {
+						stack.Path = nestedError.Path
+					}
+					if nestedError.Code != 0 {
+						stack.Code = nestedError.Code
+					}
+					if nestedError.Description != "" {
+						stack.Description = nestedError.Description
+					}
+
+					message.Stack = append(message.Stack, stack)
+				}
+			}
+		}
+	} else {
+		message.Message = err.Error()
+	}
+
+	return message
+}
+
+func (r *ApiErrorResponse) ToError() *errors.SystemError {
+	err := errors.NewWithCode(r.Message, r.Code)
+	if len(r.Stack) > 0 {
+		err.NestedError = make([]errors.NestedError, 0)
+		for _, stack := range r.Stack {
+			nestedError := errors.NestedError{
+				Message:     stack.Message,
+				Code:        stack.Code,
+				Path:        stack.Path,
+				Description: stack.Description,
+			}
+			err.NestedError = append(err.NestedError, nestedError)
+		}
+	}
+
+	return err
 }