+++ src/reverse_proxy/main.go
@@ -20,15 +20,31 @@ import (
 	data_models "github.com/Parallels/prl-devops-service/data/models"
 	"github.com/Parallels/prl-devops-service/helpers"
 	"github.com/Parallels/prl-devops-service/mappers"
+	"github.com/Parallels/prl-devops-service/reverse_proxy/models"
 	"github.com/Parallels/prl-devops-service/serviceprovider"
 )
 
 var globalReverseProxyService *ReverseProxyService
 
+type reverseProxyOperationRequest struct {
+	operation func() error
+	result    chan error
+}
+
+type reverseProxyServiceState int
+
+const (
+	ReverseProxyServiceStateStopped reverseProxyServiceState = iota
+	ReverseProxyServiceStateStarting
+	ReverseProxyServiceStateStarted
+	ReverseProxyServiceStateStopping
+)
+
 type ReverseProxyService struct {
 	enabled          bool
 	host             string
 	port             string
+	State            reverseProxyServiceState
 	forwarding_hosts []*data_models.ReverseProxyHost
 	db               *data.JsonDatabase
 	api_ctx          basecontext.ApiContext
@@ -37,6 +53,9 @@ type ReverseProxyService struct {
 	tcpListeners     []net.Listener
 	httpListeners    []*http.Server
 	wg               *sync.WaitGroup
+
+	opQueue   chan reverseProxyOperationRequest
+	queueOnce sync.Once
 }
 
 func Get(ctx basecontext.ApiContext) *ReverseProxyService {
@@ -45,6 +64,36 @@ func Get(ctx basecontext.ApiContext) *ReverseProxyService {
 	}
 
 	return globalReverseProxyService
+}
+
+func GetConfig() models.ReverseProxyConfig {
+	cfg := config.Get()
+	result := models.ReverseProxyConfig{
+		Enabled: false,
+	}
+
+	if cfg == nil || globalReverseProxyService == nil {
+		result.Enabled = false
+	}
+
+	result.Host = cfg.ReverseProxyHost()
+	result.Port = cfg.ReverseProxyPort()
+	result.Enabled = cfg.IsReverseProxyEnabled()
+
+	// Returning db data if available and not different from the config
+	dtoRp, err := globalReverseProxyService.db.GetReverseProxyConfig(globalReverseProxyService.api_ctx)
+	if dtoRp != nil && err == nil {
+		if !dtoRp.Diff(mappers.ConfigReverseProxyToDto(result)) {
+			result.Host = dtoRp.Host
+			result.Port = dtoRp.Port
+			result.Enabled = dtoRp.Enabled
+			result.ID = dtoRp.ID
+		}
+
+		return result
+	}
+
+	return result
 }
 
 func New(ctx basecontext.ApiContext) *ReverseProxyService {
@@ -60,6 +109,7 @@ func New(ctx basecontext.ApiContext) *ReverseProxyService {
 		tcpListeners:  []net.Listener{},
 		httpListeners: []*http.Server{},
 		wg:            &sync.WaitGroup{},
+		State:         ReverseProxyServiceStateStopped,
 	}
 
 	return globalReverseProxyService
@@ -117,11 +167,48 @@ func (rps *ReverseProxyService) ImportFromConfig() error {
 	return nil
 }
 
+func (rps *ReverseProxyService) CheckChangesInConfiguration() error {
+	cfg := config.Get()
+	dtoRp, _ := rps.db.GetReverseProxyConfig(rps.api_ctx)
+	if dtoRp == nil && cfg.IsReverseProxyEnabled() {
+		dto := data_models.ReverseProxy{
+			ID:      helpers.GenerateId(),
+			Enabled: cfg.IsReverseProxyEnabled(),
+			Host:    cfg.ReverseProxyHost(),
+			Port:    cfg.ReverseProxyPort(),
+		}
+		if _, err := rps.db.UpdateReverseProxy(rps.api_ctx, dto); err != nil {
+			return err
+		}
+
+		dtoRp = &dto
+		rps.host = dto.Host
+		rps.port = dto.Port
+		rps.api_ctx.LogDebugf("Reverse proxy configuration imported: %v", dtoRp)
+	} else {
+		configRp := GetConfig()
+		if dtoRp.Diff(mappers.ConfigReverseProxyToDto(configRp)) {
+			dto := mappers.ConfigReverseProxyToDto(GetConfig())
+			if _, err := rps.db.UpdateReverseProxy(rps.api_ctx, dto); err != nil {
+				return err
+			}
+		}
+		rps.host = configRp.Host
+		rps.port = configRp.Port
+	}
+
+	return nil
+}
+
 func (rps *ReverseProxyService) LoadFromDb() error {
 	dtoRp, err := rps.db.GetReverseProxyConfig(rps.api_ctx)
 	if err != nil {
 		return err
 	}
+	if dtoRp == nil {
+		return nil
+	}
+
 	rps.host = dtoRp.Host
 	rps.port = dtoRp.Port
 	hosts, err := rps.db.GetReverseProxyHosts(rps.api_ctx, "")
@@ -137,7 +224,23 @@ func (rps *ReverseProxyService) LoadFromDb() error {
 				if route.TargetVmId != "" {
 					vm, err := prl_svc.GetVm(rps.api_ctx, route.TargetVmId)
 					if err != nil || vm.InternalIpAddress == "" || vm.InternalIpAddress == "-" || vm.State != "running" {
-						rps.api_ctx.LogErrorf("Error getting vm %s for reverse proxy route: %s", route.TargetVmId, err)
+						e := ""
+						if err != nil {
+							e = err.Error()
+						}
+						if vm == nil {
+							e = "vm could not be found"
+						}
+						if vm != nil && vm.InternalIpAddress == "" {
+							e = "vm internal ip address is empty"
+						}
+						if vm != nil && vm.InternalIpAddress == "-" {
+							e = "vm internal ip address is not assigned"
+						}
+						if vm != nil && vm.State != "running" {
+							e = "vm is not running"
+						}
+						rps.api_ctx.LogErrorf("Error getting vm %s for reverse proxy route: %s", route.TargetVmId, e)
 						hostCopy.HttpRoutes[i].TargetHost = "---"
 					} else {
 						hostCopy.HttpRoutes[i].TargetHost = vm.InternalIpAddress
@@ -148,7 +251,13 @@ func (rps *ReverseProxyService) LoadFromDb() error {
 		if hostCopy.TcpRoute != nil && hostCopy.TcpRoute.TargetVmId != "" {
 			vm, err := prl_svc.GetVm(rps.api_ctx, hostCopy.TcpRoute.TargetVmId)
 			if err != nil || vm.InternalIpAddress == "" || vm.InternalIpAddress == "-" || vm.State != "running" {
-				rps.api_ctx.LogErrorf("Error getting vm %s for reverse proxy tcp route: %s", hostCopy.TcpRoute.TargetVmId, err)
+				if err != nil {
+					rps.api_ctx.LogErrorf("Error getting vm %s for reverse proxy tcp route: %s", hostCopy.TcpRoute.TargetVmId, err)
+				} else if vm == nil {
+					rps.api_ctx.LogErrorf("Error getting vm %s for reverse proxy tcp route: vm could not be found", hostCopy.TcpRoute.TargetVmId)
+				} else if vm.InternalIpAddress == "" || vm.InternalIpAddress == "-" {
+					rps.api_ctx.LogErrorf("Error getting vm %s for reverse proxy tcp route: vm internal ip address is empty", hostCopy.TcpRoute.TargetVmId)
+				}
 				hostCopy.TcpRoute.TargetHost = "---"
 			} else {
 				hostCopy.TcpRoute.TargetHost = vm.InternalIpAddress
@@ -162,6 +271,23 @@ func (rps *ReverseProxyService) LoadFromDb() error {
 }
 
 func (rps *ReverseProxyService) Start() error {
+	rps.initQueue()
+	resultChan := make(chan error)
+	rps.opQueue <- reverseProxyOperationRequest{
+		operation: rps.startInternal,
+		result:    resultChan,
+	}
+	return <-resultChan
+}
+
+func (rps *ReverseProxyService) startInternal() error {
+	if rps.State == ReverseProxyServiceStateStarted || rps.State == ReverseProxyServiceStateStarting {
+		rps.api_ctx.LogInfof("[Reverse Proxy] Reverse proxy service already started")
+		return nil
+	}
+
+	rps.State = ReverseProxyServiceStateStarting
+	cfg := config.Get()
 	rps.ctx, rps.cancelFunc = context.WithCancel(context.Background())
 	errorChan := make(chan error, 1)
 
@@ -178,14 +304,18 @@ func (rps *ReverseProxyService) Start() error {
 		rps.api_ctx.LogErrorf("Error loading reverse proxy config from db: %s", err)
 		return err
 	}
+	if err := rps.CheckChangesInConfiguration(); err != nil {
+		rps.api_ctx.LogErrorf("Error checking changes in reverse proxy config: %s", err)
+		return err
+	}
 
 	if rps.port == "" {
-		rps.port = "8080"
-		rps.api_ctx.LogWarnf("[Reverse Proxy] Port not set for reverse proxy, using default port 8080")
+		rps.port = cfg.ReverseProxyPort()
+		rps.api_ctx.LogWarnf("[Reverse Proxy] Port not set for reverse proxy, using default port", rps.port)
 	}
 	if rps.host == "" {
-		rps.host = "localhost"
-		rps.api_ctx.LogWarnf("[Reverse Proxy] Host not set for reverse proxy, using default host localhost")
+		rps.host = cfg.ReverseProxyHost()
+		rps.api_ctx.LogWarnf("[Reverse Proxy] Host not set for reverse proxy, using default host", rps.host)
 	}
 
 	rps.api_ctx.LogInfof("[Reverse Proxy] Starting reverse proxy on %s:%s", rps.host, rps.port)
@@ -197,20 +327,43 @@ func (rps *ReverseProxyService) Start() error {
 			rps.api_ctx.LogErrorf("[Reverse Proxy] Error starting reverse proxy: %s", err)
 		}
 		return err
-		// case <-rps.ctx.Done():
-		// 	rps.api_ctx.LogInfof("[Reverse Proxy] Stopping reverse proxy due to context cancellation")
-		// 	return nil
+	case <-rps.ctx.Done():
+		rps.api_ctx.LogInfof("[Reverse Proxy] Stopping reverse proxy due to context cancellation")
+		return nil
 	default:
 	}
 
+	rps.api_ctx.LogInfof("[Reverse Proxy] Reverse proxy started")
+	rps.State = ReverseProxyServiceStateStarted
 	return nil
 }
 
 func (rps *ReverseProxyService) Stop() error {
+	rps.initQueue()
+	resultChan := make(chan error)
+	rps.opQueue <- reverseProxyOperationRequest{
+		operation: rps.stopInternal,
+		result:    resultChan,
+	}
+	return <-resultChan
+}
+
+func (rps *ReverseProxyService) stopInternal() error {
+	if rps.State == ReverseProxyServiceStateStopped || rps.State == ReverseProxyServiceStateStopping {
+		rps.api_ctx.LogInfof("[Reverse Proxy] Reverse proxy service already stopped")
+		return nil
+	}
+
+	rps.State = ReverseProxyServiceStateStopping
+
 	rps.api_ctx.LogInfof("[Reverse Proxy] Stopping reverse proxy service")
-	rps.cancelFunc()
+	// fixing a possible issue with the cancel function not being called
+	if rps.cancelFunc != nil {
+		rps.cancelFunc()
+	}
 
 	for _, listener := range rps.tcpListeners {
+		// Check if the listener is already closed
 		rps.api_ctx.LogInfof("[Reverse Proxy] [TCP Route] Closing listener")
 		if err := listener.Close(); err != nil {
 			rps.api_ctx.LogErrorf("[Reverse Proxy] [TCP Route] Error closing listener: %s", err)
@@ -218,32 +371,50 @@ func (rps *ReverseProxyService) Stop() error {
 		rps.api_ctx.LogInfof("[Reverse Proxy] [TCP Route] Listener closed")
 	}
 
+	ctxShutdown, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
 	for _, server := range rps.httpListeners {
-		ctxShutdown, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
 
 		rps.api_ctx.LogInfof("[Reverse Proxy] [HTTP Route] Closing server")
 		if err := server.Shutdown(ctxShutdown); err != nil {
-			rps.api_ctx.LogErrorf("[Reverse Proxy] [HTTP Route] Error shutting down server: %s", err)
+			if err == http.ErrServerClosed {
+				rps.api_ctx.LogInfof("[Reverse Proxy] [HTTP Route] Server already closed")
+			} else {
+				rps.api_ctx.LogErrorf("[Reverse Proxy] [HTTP Route] Error shutting down server: %s", err)
+			}
 		}
 
 		rps.api_ctx.LogInfof("[Reverse Proxy] [HTTP Route] Server closed")
 	}
 
+	cancel()
+
 	rps.wg.Wait()
 
 	rps.api_ctx.LogInfof("[Reverse Proxy] Reverse proxy service stopped")
+	rps.State = ReverseProxyServiceStateStopped
 	return nil
 }
 
 func (rps *ReverseProxyService) Restart() error {
-	if err := rps.Stop(); err != nil {
+	rps.initQueue()
+	resultChan := make(chan error)
+	rps.opQueue <- reverseProxyOperationRequest{
+		operation: rps.restartInternal,
+		result:    resultChan,
+	}
+	return <-resultChan
+}
+
+func (rps *ReverseProxyService) restartInternal() error {
+	if err := rps.stopInternal(); err != nil {
 		return err
 	}
 
 	done := make(chan error, 1)
 	go func() {
-		done <- rps.Start()
+		done <- rps.startInternal()
 	}()
 
 	return <-done
@@ -407,6 +578,7 @@ func (rps *ReverseProxyService) listenHttpRoute(host *data_models.ReverseProxyHo
 			target = fmt.Sprintf("%s:%s", host.Host, host.Port)
 		}
 
+		// TODO: Add the ability to have a rewrite of the path URL and the schema
 		if strings.EqualFold(target, req.Host) {
 			for _, route := range host.HttpRoutes {
 				if route.TargetHost == "" || route.TargetHost == "---" {
@@ -536,6 +708,20 @@ func (rps *ReverseProxyService) copyWithContext(ctx context.Context, dst io.Writ
 				return
 			}
 		}
+	}
+}
+
+func (rps *ReverseProxyService) initQueue() {
+	rps.queueOnce.Do(func() {
+		rps.opQueue = make(chan reverseProxyOperationRequest, 100)
+		go rps.processQueue()
+	})
+}
+
+func (rps *ReverseProxyService) processQueue() {
+	for req := range rps.opQueue {
+		err := req.operation()
+		req.result <- err
 	}
 }